requirements.txt
streamlit
sentence-transformers
faiss-cpu
numpy
pandas
pypdf
python-docx
arabic-reshaper
python-bidi